🚀 final docker-composition created

+++ fastapi/cookies_api.py
@@ -10,63 +10,81 @@ import shutil
 import docker
 import uuid
 from multiprocessing import Manager
-from threading import Lock
 from config import settings
 import pycurl
 
-
 COMPILETIME = 10
 
+if settings.PRODUCTION:
+    data_path = "/app/data"
+else:
+    data_path = settings.FULL_DATA_PATH
 
-class ThreadSaveList():
-    """List implementation, that stores a maximum number of items."""
+
+class ContainerList():
+    #   """List implementation, that stores cookies-container-information."""
 
     def __init__(self):
-        self._list = list()
-        self._lock = Lock()
+        self._list = Manager().list()
+#        self._lock = Lock()
 
     def append(self, value):
-        with self._lock:
-            self._list.append(value)
-            return True
+        #        with self._lock:
+        self._list.append(value)
+        return True
 
     def get(self, index):
         """Return item without removal."""
-        with self._lock:
-            return self._list[index]
+#        with self._lock:
+        return self._list[index]
 
     def remove(self, index):
-        with self._lock:
-            self._list.pop(index)
+        #        with self._lock:
+        return self._list.pop(index)
+
+    def remove_first(self, timeout):
+        #        with self._lock:
+
+        if len(self._list) > 0:
+            return self._list.pop(0)
+        else:
+            return None
 
     def remove_by_uuid(self, container_uuid):
-        with self._lock:
-            for i in range(len(self._list)):
-                if self._list[i]['uuid'] == container_uuid:
-                    self._list.pop(i)
-                    break
+        #        with self._lock:
+        for i in range(len(self._list)):
+            if self._list[i]['uuid'] == container_uuid:
+                self._list.pop(i)
+                break
 
     def get_and_remove_by_uuid(self, container_uuid):
-        with self._lock:
-            for i in range(len(self._list)):
-                if self._list[i]['uuid'] == container_uuid:
-                    c = self._list.pop(i)
-                    return c
+        #        with self._lock:
+        for i in range(len(self._list)):
+            if self._list[i]['uuid'] == container_uuid:
+                return self._list.pop(i)
 
     def length(self):
-        with self._lock:
-            return len(self._list)
+        #        with self._lock:
+        return len(self._list)
+
+    def contains_uuid(self, container_uuid):
+        #        with self._lock:
+        for i in range(len(self._list)):
+            if self._list[i]['uuid'] == container_uuid:
+                return True
+        return False
 
 
 m = Manager()
-newContainers = m.Queue(maxsize=settings.MAX_CONTAINERS)
-"""Holds the infos of the new (running) cookies-containers, that are WAITING for usage."""
+newContainers = m.Queue()  # maxsize=settings.MAX_CONTAINERS)
+# newContainers = ThreadSaveList()
+# """Holds the infos of the new (running) cookies-containers, that are WAITING for usage."""
 
-runningContainers = ThreadSaveList()
+runningContainers = ContainerList()
 
 
 def writeFiles(files: List[str], uuid: str):
-    dir = os.path.join(settings.FULL_DATA_PATH, str(uuid))
+    dir = os.path.join(data_path, str(uuid))
 
     # check if dir is already existing -> delete content
     if os.path.exists(dir):
@@ -88,9 +106,9 @@ def writeFiles(files: List[str], uuid: str):
 
 
 def save_compilation_result(container_uuid: str, project_uuid: str):
-    sourcepath = os.path.join(settings.FULL_DATA_PATH, str(container_uuid))
+    sourcepath = os.path.join(data_path, str(container_uuid))
     sourcefiles = os.listdir(sourcepath)
-    destinationpath = os.path.join(settings.FULL_DATA_PATH, str(project_uuid))
+    destinationpath = os.path.join(data_path, str(project_uuid))
     Path(destinationpath).mkdir(exist_ok=True, parents=True)
 
     print(sourcepath)
@@ -107,18 +125,19 @@ def createNewContainer():
     new_uuid = uuid.uuid4()
     new_name = f"cookies-{new_uuid}"
 
-    # create directory for this uuid
+    # build and create directory for this uuid
     uuid_dir = Path(os.path.join(settings.FULL_DATA_PATH, str(new_uuid)))
-    uuid_dir.mkdir(exist_ok=True, parents=True)
+    """This variable is a parameter for the container and thus must always be a OS-path"""
+
+    create_dir = Path(os.path.join(data_path, str(new_uuid)))
+    create_dir.mkdir(exist_ok=True, parents=True)
 
     client = docker.from_env()
     nproc_limit = docker.types.Ulimit(name="nproc", soft=3700, hard=5000)
     new_container = client.containers.run(
-        'phitux/cookies', detach=True, auto_remove=True, remove=True, mem_limit="512m", name=new_name, network="schoco", ports={'8080/tcp': ('127.0.0.1', None)}, stdin_open=True, stdout=True, stderr=True, stop_signal="SIGKILL", tty=True, ulimits=[nproc_limit], user=os.getuid(), volumes=[f"{uuid_dir}:/app/tmp"])
+        'phitux/cookies', detach=True, auto_remove=True, remove=True, mem_limit="512m", name=new_name, network="schoco", ports={'8080/tcp': ('127.0.0.1', None)}, stdin_open=True, stdout=True, stderr=True, stop_signal="SIGKILL", tty=True, ulimits=[nproc_limit], user=f"{os.getuid()}:{os.getgid()}", volumes=[f"{uuid_dir}:/app/tmp"])
     # TODO place all schoco+cookies-containers in same schoco-network -> then a cookies-container can get called by its containername!!
     # -> user-defined bridge
-
-    print("new container created: " + str(new_uuid))
 
     apiclient = docker.APIClient(base_url="unix://var/run/docker.sock")
     ip = apiclient.inspect_container(new_name)[
@@ -126,42 +145,53 @@ def createNewContainer():
     port = apiclient.inspect_container(new_name)[
         'NetworkSettings']['Ports']['8080/tcp'][0]['HostPort']
 
-    return {'id': new_container.id, 'uuid': str(new_uuid), 'ip': ip, 'port': port}
+    return {'id': new_container.id, 'uuid': str(new_uuid), 'ip': ip, 'port': port, 'in_use': False}
 
 
 def refillNewContainersQueue():
     """Just fills up the Queue to the amount of containers needed.
-    Creates Containers. Runs every time when a command gets started and a container
-    is removed from the Queue."""
-    while not newContainers.qsize() + runningContainers.length() >= settings.MAX_CONTAINERS:
+    Runs every time when a command ends and a container
+    is removed from runningContainers."""
+    while (runningContainers.length() + newContainers.qsize()) < settings.MAX_CONTAINERS:
         newContainers.put(createNewContainer())
 
 
 def fillNewContainersQueue():
     """runs ONCE at start of schoco (called in main.py)"""
+    if newContainers.qsize() > 0:
+        return
+
     # remove older cookies-containers if any
     client = docker.from_env()
     containers = client.containers.list(all=True)
     for c in containers:
         if str(c.name).startswith('cookies-'):
-            # remove dir
+            print("remove container name:", str(c.name))
+            print("remove container id:", str(c.id))
 
-            dir = os.path.join(settings.FULL_DATA_PATH, c.name[8:])
-            if os.path.exists(dir):
-                shutil.rmtree(dir)
+            # remove dir
+            try:
+                dir = os.path.join(data_path, c.name[8:])
+                if os.path.exists(dir):
+                    shutil.rmtree(dir)
+            except:
+                print("couldn't remove directory:", dir)
 
             # kill container
-            if c.status == 'running':
-                c.kill()
-                # gets autoremoved since auto-removing is set to true
-            else:
-                c.remove(force=True)
+            try:
+                if c.status == 'running':
+                    c.kill()
+                    # gets autoremoved since auto-removing is set to true
+                else:
+                    c.remove(force=True)
+            except:
+                print("concurrent deletion of same container")
 
     refillNewContainersQueue()
 
 
 def prepareCompile(files: List[str]):
-    # get next container out of queue - return if no new one is available after 3 seconds.
+    # get next container out of queue - return if no new one is available after 2 seconds.
     # prepare the container and place it inside runningContainers
     try:
         c = newContainers.get(timeout=3)
@@ -177,13 +207,17 @@ def prepareCompile(files: List[str]):
     return c
 
 
-def startCompile(ip: str, port: int):
+def startCompile(uuid: str, ip: str, port: int):
     print("start compile at port: " + str(port))
 
     # pycurl to cookies-java-server "/compile"
     buffer = BytesIO()
     c = pycurl.Curl()
-    c.setopt(c.URL, f"http://localhost:{port}/compile")
+    if settings.PRODUCTION:
+        host = f"cookies-{uuid}:8080"
+    else:
+        host = f"localhost:{port}"
+    c.setopt(c.URL, f"http://{host}/compile")
     post_data = {'timeout_cpu': COMPILETIME, 'timeout_session': COMPILETIME}
 
     # Why the 7? 🤷‍♂️ Probability and trial and error...
@@ -194,7 +228,7 @@ def startCompile(ip: str, port: int):
             c.setopt(c.WRITEDATA, buffer)
             c.perform()
             break
-        except BaseException as e:
+        except:
             print("connection error: " + str(port))
             tries -= 1
             if (tries == 0):
@@ -215,7 +249,7 @@ def kill_n_create(container_uuid: str):
     runningContainers.remove_by_uuid(container_uuid)
 
     # remove container-dir
-    dir = os.path.join(settings.FULL_DATA_PATH, container_uuid)
+    dir = os.path.join(data_path, container_uuid)
     shutil.rmtree(dir)
 
     # kill and (hopefully automatically) remove container
@@ -245,14 +279,14 @@ def prepare_execute(project_uuid: str):
     try:
         c = newContainers.get(timeout=3)
     except queue.Empty:
-        return {'success': False, 'message': 'No worker ready for execution within 3 seconds 😥 Please retry!'}
+        return {'success': False, 'message': 'No worker ready for compilation within 3 seconds 😥 Please retry!'}
 
     runningContainers.append(c)
 
     # copy .class files to container-mount
-    sourcepath = os.path.join(settings.FULL_DATA_PATH, project_uuid)
+    sourcepath = os.path.join(data_path, project_uuid)
     sourcefiles = os.listdir(sourcepath)
-    destinationpath = os.path.join(settings.FULL_DATA_PATH, str(c['uuid']))
+    destinationpath = os.path.join(data_path, str(c['uuid']))
     Path(destinationpath).mkdir(exist_ok=True, parents=True)
 
     filesExist = False
@@ -265,22 +299,25 @@ def prepare_execute(project_uuid: str):
     # if no .class files existed, then return container to newContainers-Queue
     if not filesExist:
         c = runningContainers.get_and_remove_by_uuid(c['uuid'])
-        newContainers.put(c)
+        newContainers.append(c)
 
         return {'executable': False}
-
-    print(c['uuid'] + " with port " + c['port'])
 
     return c
 
 
-def start_execute(ip: str, port: int):
+def start_execute(uuid: str, ip: str, port: int):
     print("start execute at port: " + str(port))
 
     # pycurl to cookies-java-server "/execute"
     buffer = BytesIO()
     c = pycurl.Curl()
-    c.setopt(c.URL, f"http://localhost:{port}/execute")
+    if settings.PRODUCTION:
+        host = f"cookies-{uuid}:8080"
+    else:
+        host = f"localhost:{port}"
+    c.setopt(c.URL, f"http://{host}/execute")
+
     post_data = {'timeout_cpu': COMPILETIME, 'timeout_session': COMPILETIME}
 
     # Why the 7? 🤷‍♂️ Probability and trial and error...
@@ -300,18 +337,22 @@ def start_execute(ip: str, port: int):
             time.sleep(0.2)
     c.close()
 
-    print(buffer.getvalue().decode('utf-8'))
-
     return json.loads(buffer.getvalue().decode('utf-8'), strict=False)
 
 
-def start_test(ip: str, port: int):
+def start_test(uuid: str, ip: str, port: int):
     print("start test at port: " + str(port))
 
     # pycurl to cookies-java-server "/test"
     buffer = BytesIO()
     c = pycurl.Curl()
-    c.setopt(c.URL, f"http://localhost:{port}/test")
+
+    if settings.PRODUCTION:
+        host = f"cookies-{uuid}:8080"
+    else:
+        host = f"localhost:{port}"
+    c.setopt(c.URL, f"http://{host}/test")
+
     post_data = {'timeout_cpu': COMPILETIME, 'timeout_session': COMPILETIME}
 
     # Why the 7? 🤷‍♂️ Probability and trial and error...
@@ -340,14 +381,12 @@ def start_test(ip: str, port: int):
         passed_tests = int(lastLine.split('(')[1].split(')')[0].split()[0])
         results['passed_tests'] = passed_tests
         results['failed_tests'] = 0
-        print(passed_tests)
     elif lastLine.startswith('Tests run'):
         failed_tests = int(lastLine.split(',')[1].split(':')[1].strip())
         passed_tests = int(lastLine.split(',')[0].split(':')[
             1].strip()) - failed_tests
         results['passed_tests'] = passed_tests
         results['failed_tests'] = failed_tests
-        print(passed_tests, failed_tests)
     else:
         results['passed_tests'] = 0
         results['failed_tests'] = 0
